Extracting features from SentiLex corpus  ➕ - Also fixed typing in LCA feature columns

+++ models/modules_features/support/module_lca.py
@@ -118,14 +118,14 @@ def lca_analysis_dynamic(train_X: pd.DataFrame, train_Y: pd.Series, test_X: Opti
         # LCA features applied to train
         for index, cluster_center in enumerate(cluster_centers):
             column_name : str   = f'LCA - Max Cossine w/ Cluster {index}'
-            train_X[column_name] = train_X['LCA - Embedding Groups'].progress_apply(lambda sentence_embeddings: get_max_cossine_similarity_with_word(sentence_embeddings, cluster_center))
+            train_X[column_name] = train_X['LCA - Embedding Groups'].progress_apply(lambda sentence_embeddings: get_max_cossine_similarity_with_word(sentence_embeddings, cluster_center)).astype(float)
             train_X[column_name].fillna(DEFAULT_VALUE, inplace=True)
         
         # LCA features applied to test
         if test_X is not None:
             for index, cluster_center in enumerate(cluster_centers):
                 column_name : str   = f"LCA - Max Cossine w/ Cluster {index} for {cluster_to_study['full_code']}"
-                test_X[column_name] = test_X['LCA - Embedding Groups'].progress_apply(lambda sentence_embeddings: get_max_cossine_similarity_with_word(sentence_embeddings, cluster_center))
+                test_X[column_name] = test_X['LCA - Embedding Groups'].progress_apply(lambda sentence_embeddings: get_max_cossine_similarity_with_word(sentence_embeddings, cluster_center)).astype(float)
                 test_X[column_name].fillna(DEFAULT_VALUE, inplace=True)
 
     del word2vec_model
